+++ assembl/static2/js/app/utils/routeMap.js
@@ -4,7 +4,7 @@ import { stringify } from 'query-string';
 import parse from './literalStringParser';
 import { capitalize } from './globalFunctions';
 import { browserHistory } from '../router';
-import { basePath } from './csrf';
+import { basePath } from './server';
 
 /*
   A global map of routes managed by React front-end.
